fix: widen cpuUsableCores to float, restore v3.2.0 changelog, bump to v3.3.0
- cpuUsableCores is computed from float math (total_cores - reserve_cores)
  and was silently truncating fractional values. Now float.
- cpuReservedCores stays int — deliberate user-configured whole-core value.
- Restore original v3.2.0 changelog entry (CPU core fields in heartbeat).
- New v3.3.0 entry covers the system-metrics v3 / queue-metrics v3 chain.

+++ src/Manager/AutoscaleManager.php
@@ -270,7 +270,7 @@ final class AutoscaleManager
             capacityLimiter: $this->clusterString($capacity->limitingFactor, 'unknown'),
             cpuPercent: $this->clusterFloat($cpuDetails['current_cpu_percent'] ?? 0.0),
             cpuCores: is_numeric($cpuDetails['total_cores'] ?? null) ? (float) $cpuDetails['total_cores'] : 0.0,
-            cpuUsableCores: is_numeric($cpuDetails['usable_cores'] ?? null) ? (int) $cpuDetails['usable_cores'] : 0,
+            cpuUsableCores: is_numeric($cpuDetails['usable_cores'] ?? null) ? (float) $cpuDetails['usable_cores'] : 0.0,
             cpuReservedCores: is_numeric($cpuDetails['reserve_cores'] ?? null) ? (int) $cpuDetails['reserve_cores'] : 0,
             memoryPercent: $this->clusterFloat($memoryDetails['current_memory_percent'] ?? 0.0),
             memoryTotalMb: $memoryTotalMb,
